pyproject.toml
[project]
name = "projetia-devpro"
version = "0.1.0"
description = "Générateur d'images IA - MVP école"
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    # Core
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    # Queue / DB
    "redis>=5.0",
    "arq>=0.26",
    "asyncpg>=0.29",
    # HTTP clients
    "httpx>=0.27",
    # UI
    "gradio>=4.40",
    # Image
    "Pillow>=10.4",
    # Logging
    "structlog>=24.4",
    # Traduction FR→EN
    "transformers>=4.40",
    "sentencepiece>=0.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6",
    "mypy>=1.11",
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "pre-commit>=3.8",
    "ipython>=8.26",
    "types-redis",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# -----------------------------------------------------------------------------
# Ruff : lint + format
# -----------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = [
    "E501",  # line too long (géré par le formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011"]  # allow asserts

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# -----------------------------------------------------------------------------
# Mypy : type check (informatif en CI, pas bloquant)
# -----------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false

# -----------------------------------------------------------------------------
# Pytest
# -----------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
markers = [
    "integration: tests d'intégration nécessitant services externes (Postgres, Redis, ComfyUI)",
]